+++ applications/smurf/libsmurf/smurf_makemap.c
@@ -727,6 +727,7 @@
 
 #include <string.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <sys/time.h>
 #include <time.h>
 #include <math.h>
@@ -815,7 +816,7 @@ void smurf_makemap( int *status ) {
   int moving = 0;            /* Is the telescope base position changing? */
   int nparam = 0;            /* Number of extra parameters for pixel spreading*/
   size_t ntile;              /* Number of output tiles */
-  size_t nused;              /* No. of used input samples */
+  int64_t nused;             /* No. of used input samples */
   int nval;                  /* Number of parameter values supplied */
   size_t nxy;                /* Number of pixels in output image */
   smfData *odata=NULL;       /* Pointer to output SCUBA2 data struct */
